+++ src/lib/HealthReporter.ts
